Fixed display bug for card description change
If a card description changes and the previous description was empty
then the notification doesn't fire for it because the line that says
if(response.data.old.desc) that is meant to test for whether or not the
property exists, returns false since it's actually test the value of
response.data.old which is an empty string.

The correct way to do this is to check for whether or not the value is
undefined. This should ideally be done across all the other checks as
well, but since description is the only one that behaves this way
(every other card change provides some value for old) I'll be changing
it later

Bumped version to 0.1.1