Login Happens by Default on first run
Since the trello add on icon really hides at the bottom, it's a bit
confusing when the users open the app for the first time. For
onboarding experience, I shifted the process of having to click on the
icon to get started for the first time, into a default action where if
the add on doesn't have a login value stored, then it by default
opens a login page which can be dismissed and will be shown again only
wen the user reclicks the icon.

With relation to how the click to reactivaate the login window works,
some code was changed in the assigning of the trelloid variable. The
code works in the following way. When the login tab opens, the tab id is
stored in the trelloid variable. If the user clicks the icon again
while that tab is still open, a new tab isn't opened since one is
already open. If the user closes a tab without completing the login
process, this id is reset to null so that the condition for opening a
login tab becomes satisfied when the icon is next clicked.